Fixed logic with removing from personal list but not global list

+++ BYG/byg.py
@@ -223,15 +223,19 @@ def remove_from_my_list():
         ub_item = session_db.query(UserBucket).filter_by(
             id=user_bucket_id, user_netid=user_netid).first()
         if ub_item:
-            bucket_id=ub_item.bucket_id
+            bucket_id = ub_item.bucket_id
             session_db.delete(ub_item)
+            session_db.commit()
+
+            # Check if the item is public before deleting from the global list
             item = session_db.query(Bucket).filter_by(
-            bucket_id=bucket_id).first()
-            if item:
+                bucket_id=bucket_id).first()
+            if item and not item.priv:
+                # Do not delete from the global list if the item is public
+                pass
+            else:
                 session_db.delete(item)
                 session_db.commit()
-        
-        session_db.commit()
 
     return flask.redirect('/my_bucket')
